Rename instance inspector header
