+++ barretenberg/cpp/src/barretenberg/plonk_honk_shared/instance_inspector.cpp
@@ -1,3 +1,3 @@
-#include "instance_inspector.hpp"
+#include "proving_key_inspector.hpp"
 
 // Hack to make the module compile.
