+++ barretenberg/cpp/src/barretenberg/plonk_honk_shared/proving_key_inspector.hpp
