+++ barretenberg/cpp/src/barretenberg/ultra_honk/oink_prover.cpp
@@ -1,5 +1,5 @@
 #include "barretenberg/ultra_honk/oink_prover.hpp"
-#include "barretenberg/plonk_honk_shared/instance_inspector.hpp"
+#include "barretenberg/plonk_honk_shared/proving_key_inspector.hpp"
 #include "barretenberg/relations/logderiv_lookup_relation.hpp"
 
 namespace bb {
